Allow regular expressions in config file (#19)
Problem: testing the burner wallet in a local network requires to update
the configuration with the IP/hostname of the development server. My IP
changes depending on the network I'm in, and every time I get a new IP
address I have to update the config file to match it.
Moreover, if I want to test with ngrok my hostname changes all the times
I start the ngrok tunneling service.

Solution: local networks have specific IP ranges, so instead of pointing
to a specific IP address we can use a regexp to match all IPs in that
local network. We can extend this idea to all subdomains of ngrok.